JE2-143 cascade deletion & archiving added for service categories; service form updated

+++ api/src/services/service-categories.service.ts
@@ -4,11 +4,15 @@ import { ServiceCategory, ServiceCategoryDocument } from '../schemas/service-cat
 import { Model } from 'mongoose'
 import { CreateServiceCategoryDto } from '../dto/create-service-category.dto'
 import { UpdateServiceCategoryDto } from '../dto/update-service-category.dto'
+import { Service, ServiceDocument } from 'src/schemas/service.schema'
+import { ServicesService } from './services.service'
 
 @Injectable()
 export class ServiceCategoriesService {
   constructor(
     @InjectModel(ServiceCategory.name) private readonly serviceCategoryModel: Model<ServiceCategoryDocument>,
+    @InjectModel(Service.name) private readonly serviceModel: Model<ServiceDocument>,
+    private readonly servicesService: ServicesService,
   ) {}
 
   async getAll() {
@@ -22,9 +26,9 @@ export class ServiceCategoriesService {
   async getById(id: string) {
     const serviceCategory = await this.serviceCategoryModel.findById(id).exec()
 
-    if (!serviceCategory) throw new NotFoundException('Категория услуги не найдена')
+    if (!serviceCategory) throw new NotFoundException('Категория услуг не найдена')
 
-    if (serviceCategory.isArchived) throw new ForbiddenException('Категория услуги в архиве')
+    if (serviceCategory.isArchived) throw new ForbiddenException('Категория услуг в архиве')
 
     return serviceCategory
   }
@@ -32,7 +36,7 @@ export class ServiceCategoriesService {
   async getArchivedById(id: string) {
     const serviceCategory = await this.serviceCategoryModel.find({ isArchived: true }).findById(id).exec()
 
-    if (!serviceCategory) throw new NotFoundException('Категория услуги в архиве не найдена')
+    if (!serviceCategory) throw new NotFoundException('Категория услуг в архиве не найдена')
 
     return serviceCategory
   }
@@ -44,9 +48,9 @@ export class ServiceCategoriesService {
   async update(id: string, serviceCategoryDto: UpdateServiceCategoryDto, force: boolean = false) {
     const serviceCategory = await this.serviceCategoryModel.findById(id).exec()
 
-    if (!serviceCategory) throw new NotFoundException('Категория услуги не найдена')
+    if (!serviceCategory) throw new NotFoundException('Категория услуг не найдена')
 
-    if (!force && serviceCategory.isArchived) throw new ForbiddenException('Категория услуги в архиве')
+    if (!force && serviceCategory.isArchived) throw new ForbiddenException('Категория услуг в архиве')
 
     serviceCategory.set(serviceCategoryDto)
     await serviceCategory.save()
@@ -54,22 +58,40 @@ export class ServiceCategoriesService {
     return serviceCategory
   }
 
+  async isLocked(id: string) {
+    const serviceCategory = await this.serviceCategoryModel.findById(id)
+
+    if (!serviceCategory) throw new NotFoundException('Категория услуг не найдена')
+
+    const services = await this.serviceModel.find({ serviceCategory: serviceCategory._id })
+
+    if (!services.length) return false
+
+    return await Promise.any(services.map(x => this.servicesService.isLocked(String(x._id))))
+  }
+
   async archive(id: string) {
+    if (await this.isLocked(id))
+      throw new ForbiddenException('Категория услуг не может быть перемещена в архив, поскольку уже содержит услуги.',)
+
     const serviceCategory = await this.serviceCategoryModel.findById(id).exec()
 
-    if (!serviceCategory) throw new NotFoundException('Категория услуги не найдена')
+    if (!serviceCategory) throw new NotFoundException('Категория услуг не найдена')
 
-    if (serviceCategory.isArchived) throw new ForbiddenException('Категория услуги уже в архиве')
+    if (serviceCategory.isArchived) throw new ForbiddenException('Категория услуг уже в архиве')
 
     serviceCategory.isArchived = true
     await serviceCategory.save()
 
-    return { message: 'Категория услуги перемещена в архив' }
+    return { message: 'Категория услуг перемещена в архив' }
   }
 
   async delete(id: string) {
+    if (await this.isLocked(id))
+      throw new ForbiddenException('Категория услуг не может быть удалена, поскольку уже содержит услуги.')
+
     const serviceCategory = await this.serviceCategoryModel.findByIdAndDelete(id).exec()
-    if (!serviceCategory) throw new NotFoundException('Категория услуги не найдена')
-    return { message: 'Категория услуги успешно удалена' }
+    if (!serviceCategory) throw new NotFoundException('Категория услуг не найдена')
+    return { message: 'Категория услуг успешно удалена' }
   }
 }
